pyproject.toml
[project]
name = "travel_agent"
version = "0.1.0"
description = "AI travel assistant project"
authors = [
    { name = "Jaclyn Lunger", email = "lunger.jaclyn@gmail.com" }
]

dependencies = [
    "amadeus==11.0.0",
    "annotated-types==0.7.0",
    "anyio==4.8.0",
    "certifi==2024.12.14",
    "click==8.1.8",
    "distro==1.9.0",
    "fastapi==0.115.6",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httpx==0.28.1",
    "idna==3.10",
    "jiter==0.8.2",
    "openai==1.59.8",
    "pydantic==2.10.5",
    "pydantic-core==2.27.2",
    "python-dotenv==1.0.1",
    "sniffio==1.3.1",
    "starlette==0.41.3",
    "tqdm==4.67.1",
    "typing-extensions==4.12.2",
    "uvicorn==0.34.0",
    "charset-normalizer==3.4.1",
    "requests==2.32.3",
    "urllib3==2.3.0",
    "django-admin==2.0.2",
    "django-excel-response2==3.0.6",
    "django-six==1.0.5",
    "excel-base==1.0.4",
    "isoweek==1.3.3",
    "python-dateutil==2.9.0",
    "screen==1.0.1",
    "six==1.17.0",
    "timeconvert==3.0.13",
    "tzlocal==5.2",
    "xlwt==1.3.0",
    "asgiref==3.8.1",
    "django==5.1.5",
    "sqlparse==0.5.3",
    "django-cors-headers==4.6.0",
    "openai==1.59.8",
    "django-extensions==3.2.3",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["utils*", "travel_agent*"]

